+++ skrub/_duration_to_float.py
