+++ skrub/tests/test_duration_to_float.py
@@ -5,11 +5,11 @@ import pytest
 from numpy.testing import assert_allclose
 
 import skrub._dataframe as sbd
-from skrub._convert_duration import DurationToFloat, duration_to_float
+from skrub._duration_to_float import DurationToFloat, duration_to_float
 from skrub.core import RejectColumn
 
 
-def test_convert_duration(df_module):
+def test_duration_to_float(df_module):
     df = df_module.make_column(
         "duration",
         [
@@ -27,14 +27,14 @@ def test_convert_duration(df_module):
     assert_allclose(sbd.to_numpy(transformed), sbd.to_numpy(expected))
 
 
-def test_convert_duration_rejects_non_duration(df_module):
+def test_duration_to_float_rejects_non_duration(df_module):
     df = df_module.make_column("not_duration", [1, 2, 3, 4])
     transformer = DurationToFloat()
     with pytest.raises(RejectColumn, match="Expected a duration column, got*"):
         transformer.fit_transform(df)
 
 
-def test_dispatched_convert_duration(df_module):
+def test_dispatched_duration_to_float(df_module):
     s = df_module.make_column(
         "", [timedelta(days=1), timedelta(hours=1), timedelta(microseconds=1)]
     )
